feat(agents): add LLM-guided edit page for existing agents
Add /agents/$id/edit route with a chat + YAML editor split layout for
LLM-assisted editing of an existing agent's config, mirroring the new
agent creation flow. The agent detail page becomes a layout route with
an index child; its Edit button keeps the YAML-only modal, and the
actions menu gains a "Guided edit" item that routes to the new page.

Widen the edit modal (max-w-3xl, 85dvh) so longer configs fit, and
keep the CodeEditor border visible by clipping overflow on the inner
editor rather than the bordered wrapper.
