+++ apps/dashboard/src/client/ui/routes/agents/$id/edit.tsx
@@ -1,0 +1,184 @@
+import { useEffect, useState, type ReactNode } from "react";
+import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
+import { createFileRoute, useNavigate } from "@tanstack/react-router";
+import { ArrowLeft } from "lucide-react";
+import { parse, stringify } from "yaml";
+import { toast } from "sonner";
+
+import { Button } from "@hermeum/components/ui/button";
+import {
+  ResizableHandle,
+  ResizablePanel,
+  ResizablePanelGroup,
+} from "@hermeum/components/ui/resizable";
+import { useTRPC } from "@/router";
+import { AgentInputObjectSchema, AgentInputSchema } from "@/entities";
+import type { AgentInput } from "@/entities";
+import { AgentConfigChat } from "@/client/ui/components/agent-config-chat";
+import { CodeEditor } from "@/client/ui/components/code-editor";
+
+const YAML_OPTIONS = { blockQuote: "literal", lineWidth: 0 } as const;
+
+export const Route = createFileRoute("/agents/$id/edit")({
+  component: EditAgentPage,
+});
+
+function agentToYaml(agent: unknown): string {
+  return stringify(AgentInputObjectSchema.parse(agent), YAML_OPTIONS).trim();
+}
+
+function EditAgentPage() {
+  const { id } = Route.useParams();
+  const trpc = useTRPC();
+  const queryClient = useQueryClient();
+  const navigate = useNavigate();
+  const [editorValue, setEditorValue] = useState("");
+  const [validationError, setValidationError] = useState<string | null>(null);
+  const [seeded, setSeeded] = useState(false);
+
+  const { data: agent, isPending, error } = useQuery(
+    trpc.agent.get.queryOptions({ id })
+  );
+
+  // Seed the editor with the existing agent's config once it loads. The draft
+  // is what the chat LLM reads/writes, so editing starts from the persisted
+  // definition rather than a blank slate.
+  useEffect(() => {
+    if (agent && !seeded) {
+      setEditorValue(agentToYaml(agent));
+      setSeeded(true);
+    }
+  }, [agent, seeded]);
+
+  const { mutate: updateAgent, isPending: isUpdating } = useMutation(
+    trpc.agent.update.mutationOptions({
+      onSuccess: (updated) => {
+        toast.success("Agent updated");
+        queryClient.setQueryData(trpc.agent.get.queryKey({ id }), updated);
+        queryClient.invalidateQueries({ queryKey: trpc.agent.list.queryKey() });
+        navigate({ to: "/agents/$id", params: { id } });
+      },
+      onError: (error) => {
+        setValidationError(error.message);
+      },
+    })
+  );
+
+  function getConfig(): AgentInput | undefined {
+    let parsed: unknown;
+    try {
+      parsed = parse(editorValue) ?? {};
+    } catch {
+      return undefined;
+    }
+    const result = AgentInputObjectSchema.safeParse(parsed);
+    return result.success ? result.data : undefined;
+  }
+
+  function handleConfigUpdate(config: AgentInput) {
+    setEditorValue(stringify(config, YAML_OPTIONS).trim());
+    setValidationError(null);
+  }
+
+  function handleSave() {
+    let parsed: Record<string, unknown>;
+    try {
+      parsed = (parse(editorValue) ?? {}) as Record<string, unknown>;
+    } catch (e) {
+      setValidationError(e instanceof Error ? e.message : "Invalid YAML");
+      return;
+    }
+
+    const result = AgentInputSchema.safeParse(parsed);
+    if (!result.success) {
+      const issue = result.error.issues[0]!;
+      const path = issue.path.length > 0 ? `/${issue.path.join("/")}` : "";
+      setValidationError(`${issue.message} (path: ${path})`);
+      return;
+    }
+
+    setValidationError(null);
+    updateAgent({ id, ...parsed });
+  }
+
+  if (isPending) return <div className="p-6">Loading…</div>;
+  if (error) return <div className="p-6 text-red-500">Error: {error.message}</div>;
+  if (!agent) return <div className="p-6">Not found</div>;
+
+  const configPane: ReactNode = (
+    <div className="flex min-h-0 flex-1 flex-col gap-2">
+      <p className="shrink-0 text-sm font-medium">Agent config</p>
+      <div className="min-h-0 flex-1">
+        <CodeEditor
+          value={editorValue}
+          onChange={setEditorValue}
+          invalid={!!validationError}
+          height="100%"
+        />
+      </div>
+      {validationError && (
+        <p className="shrink-0 text-sm text-destructive">{validationError}</p>
+      )}
+      <div className="flex shrink-0 justify-end gap-2">
+        <Button
+          variant="outline"
+          onClick={() => navigate({ to: "/agents/$id", params: { id } })}
+        >
+          Cancel
+        </Button>
+        <Button onClick={handleSave} disabled={isUpdating}>
+          {isUpdating ? "Saving…" : "Save changes"}
+        </Button>
+      </div>
+    </div>
+  );
+
+  return (
+    <div className="flex h-full min-h-0 flex-col gap-4 p-6">
+      <div className="flex shrink-0 items-center gap-1">
+        <Button
+          variant="ghost"
+          size="icon-xs"
+          aria-label="Back to agent"
+          onClick={() => navigate({ to: "/agents/$id", params: { id } })}
+        >
+          <ArrowLeft />
+        </Button>
+        <h1 className="text-2xl font-semibold tracking-tight">
+          Edit agent
+          <span className="text-muted-foreground"> · {agent.name ?? agent.id}</span>
+        </h1>
+      </div>
+
+      {/* Mobile: static stacked layout (config on top, chat at bottom) */}
+      <div className="flex min-h-0 flex-1 flex-col gap-6 lg:hidden">
+        <div className="flex min-h-0 flex-1 flex-col">{configPane}</div>
+        <div className="flex min-h-0 flex-1 flex-col">
+          <AgentConfigChat getConfig={getConfig} onConfigUpdate={handleConfigUpdate} />
+        </div>
+      </div>
+
+      {/* Desktop: resizable side-by-side layout */}
+      <ResizablePanelGroup
+        orientation="horizontal"
+        className="hidden min-h-0 flex-1 lg:flex"
+      >
+        {/* Chat pane */}
+        <ResizablePanel defaultSize="50%" minSize="25%" className="min-h-0">
+          <div className="flex h-full min-h-0 flex-col pr-3">
+            <AgentConfigChat getConfig={getConfig} onConfigUpdate={handleConfigUpdate} />
+          </div>
+        </ResizablePanel>
+        <ResizableHandle
+          withHandle
+          className="bg-transparent [&>div]:opacity-0 [&>div]:transition-opacity hover:[&>div]:opacity-100"
+        />
+
+        {/* Config pane */}
+        <ResizablePanel defaultSize="50%" minSize="25%" className="min-h-0">
+          <div className="flex h-full min-h-0 flex-col pl-3">{configPane}</div>
+        </ResizablePanel>
+      </ResizablePanelGroup>
+    </div>
+  );
+}
